# feat: only save session to storage when it changes