--- a/lib/fastifySession.js
+++ b/lib/fastifySession.js
@@ -200,7 +200,7 @@ function isConnectionEncrypted (request) {
 }
 
 function shouldSaveSession (request, cookieOpts, saveUninitialized) {
-  if (!saveUninitialized && !isSessionModified(request.session)) {
+  if (!saveUninitialized && !request.session.isModified()) {
     return false
   }
   if (cookieOpts.secure !== true || cookieOpts.secure === 'auto') {
@@ -211,10 +211,6 @@ function shouldSaveSession (request, cookieOpts, saveUninitialized) {
   }
   const forwardedProto = getRequestProto(request)
   return forwardedProto === 'https'
-}
-
-function isSessionModified (session) {
-  return (Object.keys(session).length !== 4)
 }
 
 function option (options, key, def) {