--- a/lib/session.js
+++ b/lib/session.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const crypto = require('crypto')
+
 const Cookie = require('./cookie')
 const cookieSignature = require('cookie-signature')
 
@@ -10,6 +12,8 @@ const addDataToSession = Symbol('addDataToSession')
 const generateId = Symbol('generateId')
 const requestKey = Symbol('request')
 const cookieOptsKey = Symbol('cookieOpts')
+const originalHash = Symbol('originalHash')
+const hash = Symbol('hash')
 
 module.exports = class Session {
   constructor (request, idGenerator, cookieOpts, secret, prevSession = {}) {
@@ -26,6 +30,7 @@ module.exports = class Session {
       this.sessionId = this[generateId](this[requestKey])
       this.encryptedSessionId = this[sign]()
     }
+    this[originalHash] = this[hash]()
   }
 
   touch () {
@@ -143,12 +148,34 @@ module.exports = class Session {
     return cookieSignature.sign(this.sessionId, this[secretKey])
   }
 
+  [hash] () {
+    const sess = this
+    const str = JSON.stringify(sess, function (key, val) {
+      // ignore sess.cookie property
+      if (this === sess && key === 'cookie') {
+        return
+      }
+
+      return val
+    })
+
+    return crypto
+      .createHash('sha1')
+      .update(str, 'utf8')
+      .digest('hex')
+  }
+
+  isModified () {
+    return this[originalHash] !== this[hash]()
+  }
+
   static restore (request, idGenerator, cookieOpts, secret, prevSession) {
     const restoredSession = new Session(request, idGenerator, cookieOpts, secret, prevSession)
     const restoredCookie = new Cookie(cookieOpts)
     restoredCookie.expires = new Date(prevSession.cookie.expires)
     restoredSession.cookie = restoredCookie
     restoredSession.expires = restoredCookie.expires
+    restoredSession[originalHash] = restoredSession[hash]()
     return restoredSession
   }
 }