--- a/lib/session.js
+++ b/lib/session.js
@@ -4,6 +4,9 @@ const crypto = require('crypto')
 
 const Cookie = require('./cookie')
 const cookieSignature = require('cookie-signature')
+const { configure: configureStringifier } = require('safe-stable-stringify')
+
+const stringify = configureStringifier({ bigint: false })
 
 const maxAge = Symbol('maxAge')
 const secretKey = Symbol('secretKey')
@@ -150,7 +153,7 @@ module.exports = class Session {
 
   [hash] () {
     const sess = this
-    const str = JSON.stringify(sess, function (key, val) {
+    const str = stringify(sess, function (key, val) {
       // ignore sess.cookie property
       if (this === sess && key === 'cookie') {
         return