--- a/package.json
+++ b/package.json
@@ -35,6 +35,7 @@
     "fastify": "^3.0.0",
     "fastify-cookie": "^5.3.1",
     "nyc": "^15.0.0",
+    "sinon": "^13.0.1",
     "standard": "^16.0.1",
     "tsd": "^0.19.0",
     "typescript": "^4.0.2",