--- a/package.json
+++ b/package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "cookie-signature": "^1.1.0",
     "fastify-plugin": "^3.0.0",
+    "safe-stable-stringify": "^2.3.1",
     "uid-safe": "^2.1.5"
   },
   "engines": {
@@ -40,7 +41,6 @@
     "nyc": "^15.0.0",
     "published-session": "npm:@fastify/session@latest",
     "redis": "^4.0.6",
-    "safe-stable-stringify": "^2.3.1",
     "session-file-store": "^1.5.0",
     "sinon": "^13.0.1",
     "standard": "^16.0.1",