@jsumners 2022-03-30
Would it be more performant to extend this approach to inspect the value for a change before marking the session as changed? https://github.com/jsumners/fastify-server-session/blob/f02ac8dd30dd76705b01293a749f22e73e4649b4/lib/session.js#L7-L28