@SimenB 2022-04-05
@mcollina I'd love to add a benchmark. I have one locally, but I'm not sure how realistic it is. Its results show that using the default store is faster than the approach in this, but _nobody_ should ever use the default store (memory store) in production, so that doesn't really prove anything. Using the [file store](https://www.npmjs.com/package/session-file-store) shows about a 4x improvement in performance using this branch and a redis store (probably most realistic of the three, and what we use at work) about 50%.

Of course calling the server 25 times in a tight loop is not realistic, but maybe this illustrates _something_? 😅

Difference comes from the fact we in this PR swap out `sessionStore.save` _and_ `reply.setCookie` with a single `hash` call (unless the session actually changes, which it does every 5 requests made in my benchmark).

Bench code:

```
╔═══════════════════════╤═════════╤═══════════════╤═══════════╗
║ Slower tests          │ Samples │        Result │ Tolerance ║
╟───────────────────────┼─────────┼───────────────┼───────────╢
║ old code (file)       │      24 │   1.52 op/sec │  ± 0.91 % ║
║ current code (file)   │      25 │   5.95 op/sec │  ± 3.17 % ║
║ old code (redis)      │      25 │  24.68 op/sec │  ± 1.45 % ║
║ current code (redis)  │      25 │  33.90 op/sec │  ± 2.11 % ║
║ current code (memory) │      25 │ 245.25 op/sec │  ± 3.73 % ║
╟───────────────────────┼─────────┼───────────────┼───────────╢
║ Fastest test          │ Samples │        Result │ Tolerance ║
╟───────────────────────┼─────────┼───────────────┼───────────╢
║ old code (memory)     │      25 │ 268.14 op/sec │  ± 3.90 % ║
╚═══════════════════════╧═════════╧═══════════════╧═══════════╝
```

```diff
diff --git c/bench.js w/bench.js
new file mode 100644
index 0000000..1e3ed21
--- /dev/null
+++ w/bench.js
@@ -0,0 +1,87 @@
+'use strict'
+
+const cronometro = require('cronometro')
+const Fastify = require('fastify')
+const fastifyCookie = require('fastify-cookie')
+const fileStoreFactory = require('session-file-store')
+const redisStoreFactory = require('connect-redis')
+const { createClient: createRedisClient } = require('redis')
+const fastifySession = require('.')
+const oldFastifySession = require('old-session')
+const fastifyPlugin = require('fastify-plugin')
+
+const sessionPlugin = async (fastify, { plugin, storeType }) => {
+  const FileStore = fileStoreFactory(plugin)
+  const RedisStore = redisStoreFactory(plugin)
+  const redisClient = createRedisClient({ legacyMode: true })
+  // only run this if you have redis running on localhost:6379
+  await redisClient.connect()
+
+  fastify.register(fastifyCookie)
+
+  fastify.register(plugin, {
+    secret: 'cNaoPYAwF60HZJzkcNaoPYAwF60HZJzk',
+    saveUninitialized: false,
+    cookie: { secure: false },
+    store: storeType === 'redis'
+      ? new RedisStore({ client: redisClient })
+      : storeType === 'file'
+        ? new FileStore({})
+        : undefined
+  })
+}
+
+const getServer = (plugin, storeType) => {
+  let requestCounter = 0
+
+  const fastify = Fastify()
+
+  fastify.register(fastifyPlugin(sessionPlugin), { plugin, storeType })
+
+  fastify.get('/', (request, reply) => {
+    // modify session every 10 requests
+    if (requestCounter % 10 === 0) {
+      request.session.userId = requestCounter
+    } else {
+      request.session.userId = 0
+    }
+
+    requestCounter++
+
+    reply.send(200)
+  })
+
+  return fastify
+}
+
+const testFunction = (sessionPlugin, storeType) => {
+  const server = getServer(sessionPlugin, storeType)
+
+  return async () => {
+    const { headers } = await server.inject('/')
+    const setCookieHeader = headers['set-cookie']
+
+    if (!setCookieHeader) {
+      throw new Error('Missing set-cookie header')
+    }
+
+    const { sessionId } = server.parseCookie(setCookieHeader)
+
+    // make 25 "requests" with the new session
+    for (let i = 0; i < 25; i++) {
+      await server.inject({ path: '/', headers: { cookie: `sessionId=${sessionId}` } })
+    }
+  }
+}
+
+cronometro(
+  {
+    'current code (memory)': testFunction(fastifySession),
+    'old code (memory)': testFunction(oldFastifySession),
+    'current code (file)': testFunction(fastifySession, 'file'),
+    'old code (file)': testFunction(oldFastifySession, 'file'),
+    'current code (redis)': testFunction(fastifySession, 'redis'),
+    'old code (redis)': testFunction(oldFastifySession, 'redis')
+  },
+  { iterations: 25 }
+)
diff --git c/package.json w/package.json
index 2a05309..9f90ca7 100644
--- c/package.json
+++ w/package.json
@@ -32,9 +32,14 @@
   "devDependencies": {
     "@types/node": "^17.0.0",
     "ava": "^4.1.0",
+    "connect-redis": "^6.1.3",
+    "cronometro": "^0.8.0",
     "fastify": "^3.0.0",
     "fastify-cookie": "^5.3.1",
     "nyc": "^15.0.0",
+    "old-session": "npm:@fastify/session@^8.0.0",
+    "redis": "^4.0.6",
+    "session-file-store": "^1.5.0",
     "sinon": "^13.0.1",
     "standard": "^16.0.1",
     "tsd": "^0.19.0",
```