@mcollina 2022-04-05
```
+
+    // make 25 "requests" with the new session
+    for (let i = 0; i < 25; i++) {
+      await server.inject({ path: '/', headers: { cookie: `sessionId=${sessionId}` } })
+    }
+  }
```

This benchmark code is a bit flawed when using any I/O storage. You are essentially measuring Redis Round-Trip-Time. It's significantly better to:

1. run these in parallel, e.g. `Promise.all`
2. enable autoPipelining in ioredis.

I suspect that hashing would quickly become the bottleneck (as it did for etags).